discogs-api/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-discogs-client==2.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
google-generativeai==0.8.3
